+++ changelog.md
@@ -1,3 +1,4 @@
 Version 1.0.0
 
 - Redirect to the master survey when the `Completed` question is not set
+- Send data to the master survey so it's possible to complete the current survey